updated file

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: lramela <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 12:33:56 by lramela           #+#    #+#             */
-/*   Updated: 2019/05/27 12:50:16 by lramela          ###   ########.fr       */
+/*   Updated: 2019/06/04 14:49:21 by lramela          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char		*ft_strncat(char *s1, const char *s2 , size_t n)
 {
-	int i;
-	int k;
+	int 	i;
+	size_t 	k;
 
 	i = 0;
 	k = 0;
